requirements.txt
django==3.2.8
asgiref==3.4.1
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pyjwt==1.7.1